feat: move init to lib.rs + tests
