+++ .github/workflows/merge.yaml
@@ -143,7 +143,7 @@ jobs:
           cachix_auth_token: ${{ secrets.CACHIX_AUTH_TOKEN }}
           command: |
             set -e
-            RUSTDOCFLAGS="-D warnings" nix develop -c cargo doc --no-deps --document-private-items
+            RUSTDOCFLAGS="-D warnings" nix develop -c cargo doc --no-deps --document-private-items --no-default-features --features runtime-tokio,telemetry,transport-quic
             nix build -L .#docs
       - name: Increase git buffer size
         run: git config --global http.postBuffer 524288000
